Create gene_prediction_comparison.py
ORForise is a tool used for comparison of Gene Prediction results from two different tools. This code takes the following input:
gene prediction results files from Prodigal and FragGeneScan
It gives the following output:
csv file containing the comparison of the tools based on various metrics. 

Here is the link to GitHub of ORForise: https://github.com/NickJD/ORForise


As this code required to be implemented locally, I hardcoded the file paths.